Add Iceberg support (#68)
* support kinesis firehose, unify kinesis config, add better metric support

* clean up warnings

* fix registerGauge declaration and change deprecated setDeliverGuarantee method call

* added test for firehose config

* initial version

* compileable

* close to testable

* writing tests

* Adding parallelism to sink

* got basic test working

* fix bug in test

* improve tests

* support consuming from multiple streams

* add tests

* reformat/optimize imports

* refactored several things

* fix typo

* refactor addSink() method in cassandra sink

* try to fix cluster builder serialization issues

* refactor closures in JdbcSink and CassandraSink

* fix type and reformat

* turn off debugging in tests

* updates

* updates

* add print sink and iceberg tests

* refactor iceberg source

* fix iceberg tests

* update iceberg version to 1.2.0

* add aws.region setting for tests

* remove aws.region setting for tests

* set aws creds for iceberg test

* update iceberg direct access test

* try to fix test on travis

* remove testcontainers debug from tests

* fixed test

* remove duplicative trait extension

* fix id64 test

* add simple stream node inspection for testing

* fix confluent avro deserialization

* optimize imports/scalafmt

* reenable ConfluentAvroRegistryKafkaRecordDeserializationSchemaTest

* fix tests

---------

+++ src/main/scala/io/epiphanous/flinkrunner/model/Id64.scala
@@ -66,11 +66,11 @@ object Id64 {
   def ticksOf(id: String): Long      = uuidOf(id).timestamp()
   def microsOf(id: String): Long     =
     Math
-      .floor((ticksOf(id) - GREGORIAN_OFFSET) / 10)
+      .floor((ticksOf(id) - GREGORIAN_OFFSET).toDouble / 10.0)
       .toLong
   def millisOf(id: String): Long     =
     Math
-      .floor((ticksOf(id) - GREGORIAN_OFFSET) / 10000)
+      .floor((ticksOf(id) - GREGORIAN_OFFSET).toDouble / 10000.0)
       .toLong
   def instantOf(id: String): Instant = {
     val t     = ticksOf(id) - GREGORIAN_OFFSET
